refactor: go back to monolingual tasks
